100/100

+++ cours/week_4/Длина отрезка/solution.py
@@ -1,8 +1,8 @@
 from math import sqrt, pow
-x1 = int(input())
-y1 = int(input())
-x2 = int(input())
-y2 = int(input())
+x1 = float(input())
+y1 = float(input())
+x2 = float(input())
+y2 = float(input())
 
 
 def different(a, b):
